[BRANCH] flash savings and ram savings
Flash saving with PREVENT_ARITHMETIC_OVERFLOW about 92bytes, and 32bytes RAM

This change is experimental and currently not topic to merge.
I do not really understand why chans[] is 32bits. It's not really used.
When it's actually read back for a mixloop, the resolution is immediately removed.
applyLimits does the same. 
There is only one exception: If output limits are set to -125 and +125 and an offset is also nearly e.g. -125; we would have a output range from -1024 to 1024 just for positive input values. Therefore we would need in this case about 2* the usual resolution. But there is no reason actually to have it 256* or 100* as it was originally. Also I think this is not a usual case, but also do not know if in this case a resolution drop is acceptable.

This change calculates inside mix with 32bit as it was originally, but before setting to chans[] doing >>7. So now range is -2048 to 2048, not to harm this situation limits are chosen extremely. If we would ignore this exception (it's my assumption, may be wrong) there would be even a higher flash saving, because >>8 is less effort as >>7.

I am also experimenting with ex_chans but can't get rid of it, because it would change mixloop meaning and conflict with fade phase flight. Also tried to combine sum_chans with chans but not possible either. Maybe later...